+++ re2/testing/re2_test.cc
@@ -519,7 +519,7 @@ TEST(EmptyCharset, Fuzz) {
     "[^\\D\\d]",
     "[^\\D[:digit:]]"
   };
-  for (int i = 0; i < arraysize(empties); i++)
+  for (size_t i = 0; i < arraysize(empties); i++)
     ASSERT_FALSE(RE2(empties[i]).Match("abc", 0, 3, RE2::UNANCHORED, NULL, 0));
 }
 
@@ -534,7 +534,7 @@ TEST(EmptyCharset, BitstateAssumptions) {
     "((((()))))" "(([^\\S\\s]|[^\\S\\s])|)"
   };
   StringPiece group[6];
-  for (int i = 0; i < arraysize(nop_empties); i++)
+  for (size_t i = 0; i < arraysize(nop_empties); i++)
     ASSERT_TRUE(RE2(nop_empties[i]).Match("", 0, 0, RE2::UNANCHORED, group, 6));
 }
 
@@ -1297,7 +1297,7 @@ static struct ErrorTest {
   { "zz(?P<name\377>abc)", "" },
 };
 TEST(RE2, ErrorArgs) {
-  for (int i = 0; i < arraysize(error_tests); i++) {
+  for (size_t i = 0; i < arraysize(error_tests); i++) {
     RE2 re(error_tests[i].regexp, RE2::Quiet);
     EXPECT_FALSE(re.ok());
     EXPECT_EQ(re.error_arg(), error_tests[i].error) << re.error();
@@ -1319,7 +1319,7 @@ static struct NeverTest {
 TEST(RE2, NeverNewline) {
   RE2::Options opt;
   opt.set_never_nl(true);
-  for (int i = 0; i < arraysize(never_tests); i++) {
+  for (size_t i = 0; i < arraysize(never_tests); i++) {
     const NeverTest& t = never_tests[i];
     RE2 re(t.regexp, opt);
     if (t.match == NULL) {
@@ -1454,18 +1454,18 @@ TEST(RE2, NullVsEmptyStringSubmatches) {
   // matches[0] is overall match, [1] is (), [2] is (foo), [3] is nonexistent.
   StringPiece matches[4];
 
-  for (int i = 0; i < arraysize(matches); i++)
+  for (size_t i = 0; i < arraysize(matches); i++)
     matches[i] = "bar";
 
   StringPiece null;
   EXPECT_TRUE(re.Match(null, 0, null.size(), RE2::UNANCHORED,
                        matches, arraysize(matches)));
-  for (int i = 0; i < arraysize(matches); i++) {
+  for (size_t i = 0; i < arraysize(matches); i++) {
     EXPECT_TRUE(matches[i].data() == NULL);  // always null
     EXPECT_TRUE(matches[i].empty());
   }
 
-  for (int i = 0; i < arraysize(matches); i++)
+  for (size_t i = 0; i < arraysize(matches); i++)
     matches[i] = "bar";
 
   StringPiece empty("");
